+++ horovod/spark/task/mpirun_exec_fn.py
@@ -18,7 +18,9 @@ import sys
 import threading
 import time
 
-from horovod.spark import codec, driver_service, task_service, secret
+from horovod.spark.task import task_service
+from horovod.spark.util import codec, secret
+from horovod.spark.driver import driver_service
 
 
 def parent_process_monitor(initial_ppid):
